measures not saved on sonar equal or greater than 4.4
The measures of 'basic metrics' (ncloc, files, etc.), 'complexity'
and coverage are not saved on Sonar version equal or greater than
4.4 but the source files are saved. On Sonar 3.7.4 these metrics
measures are saved.

The fix is to use a org.sonar.api.resources.File as the resource
on the sensorContext.saveMeasure method. The idea came from
looking at the source of Javascript Sonar Plugin. With this the
measures are saved in Sonar 3.7.4 and Sonar 4.4 or greater.